Removing complex objects from the AST.
Otherwise, we'd end up with string representations of Ruby objects in the AST (e.g. "<Namespace::Class:object_id>"). Front-end code trying to grok that AST is only going to be confused by such things.

Is this the right solution? Quite possibly not, but at the very least it's a start.